Cleanup structure, moving test folder to top level in preparation for sync wrapper implementation
