+++ test/api/test.sh
@@ -1,6 +1,6 @@
-python3 -m venv venv
+$PYTHON_VERSION -m venv venv
 source venv/bin/activate
 pip install -r requirements.txt
-cd athera/test
+export PYTHONPATH=$PYTHONPATH:$(pwd)/../..
 nose2 -v
 deactivate
